Moved checker to Component/Payment
