+++ src/Sylius/Bundle/CoreBundle/Resources/config/services/checkers.xml
@@ -41,6 +41,9 @@
         >
             <tag name="sylius.promotion_coupon_eligibility_checker" />
         </service>
+        <service id="sylius.checker.order_payment_method_eligibility" class="Sylius\Component\Payment\Checker\OrderPaymentMethodEligibilityChecker">
+            <argument type="service" id="sylius.context.channel" />
+        </service>
         <service id="sylius.checker.promotion_coupon.channel_eligibility" alias="sylius.promotion_coupon_channels_eligibility_checker" />
     </services>
 </container>
