+++ src/Sylius/Bundle/CoreBundle/Resources/config/services/validators.xml
@@ -65,6 +65,7 @@
         </service>
         <service id="sylius.validator.order_shipping_method_eligibility" alias="sylius.validator.shipping_method_integrity" />
         <service id="sylius.validator.payment_method_integrity" class="Sylius\Bundle\CoreBundle\Validator\Constraints\OrderPaymentMethodEligibilityValidator">
+            <argument type="service" id="sylius.checker.order_payment_method_eligibility" />
             <tag name="validator.constraint_validator" alias="sylius_order_payment_method_eligibility_validator" />
         </service>
         <service id="sylius.validator.order_payment_method_eligibility" alias="sylius.validator.payment_method_integrity" />
