+++ src/Sylius/Bundle/CoreBundle/Validator/Constraints/OrderPaymentMethodEligibilityValidator.php
@@ -13,16 +13,16 @@ declare(strict_types=1);
 
 namespace Sylius\Bundle\CoreBundle\Validator\Constraints;
 
-use Sylius\Bundle\CoreBundle\Resolver\OrderPaymentMethodEligibilityResolverInterface;
 use Sylius\Component\Core\Model\OrderInterface;
 use Sylius\Component\Core\Model\PaymentMethodInterface;
+use Sylius\Component\Payment\Checker\OrderPaymentMethodEligibilityCheckerInterface;
 use Symfony\Component\Validator\Constraint;
 use Symfony\Component\Validator\ConstraintValidator;
 use Webmozart\Assert\Assert;
 
 final class OrderPaymentMethodEligibilityValidator extends ConstraintValidator
 {
-    public function __construct(private readonly OrderPaymentMethodEligibilityResolverInterface $checker)
+    public function __construct(private readonly OrderPaymentMethodEligibilityCheckerInterface $checker)
     {
     }
 
@@ -41,7 +41,7 @@ final class OrderPaymentMethodEligibilityValidator extends ConstraintValidator
 
         foreach ($payments as $payment) {
             $paymentMethod = $payment->getMethod();
-            if ($paymentMethod instanceof PaymentMethodInterface && !$this->checker->isPaymentMethodAvailable($paymentMethod)) {
+            if ($paymentMethod instanceof PaymentMethodInterface && !$this->checker->isEligible($paymentMethod)) {
                 $this->context->addViolation(
                     $constraint->message,
                     ['%paymentMethodName%' => $paymentMethod->getName()],
